Major rewrite. Replaced rubysdl with my own C extension.
This goes from:

graphics-1.0.0b1 being 385 lines of ruby
+ rubysdl being 645 lines of ruby + 4500 lines of C (not counting deps)

to:

graphics-1.0.0b2 being 455 lines of ruby + 748 lines of C.

This unencumbers me from rubysdl, which was LGPL and hasn't had a commit
in over a year. It doesn't wrap much of SDL that I'd like to see, so
this frees me to do this much more easily.